requirements.txt
fastapi==0.115.4
uvicorn==0.32.0
pydantic
cachetools
python-multipart
Pillow